♻️ refactor(todolist): use packlets
